+++ filament/backend/src/vulkan/VulkanSwapChain.h
@@ -1,0 +1,62 @@
+/*
+ * Copyright (C) 2021 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+ #ifndef TNT_FILAMENT_DRIVER_VULKANSWAPCHAIN_H
+ #define TNT_FILAMENT_DRIVER_VULKANSWAPCHAIN_H
+
+#include "VulkanContext.h"
+#include "VulkanDriver.h"
+
+namespace filament {
+namespace backend {
+
+struct VulkanSwapChain : public HwSwapChain {
+    VulkanSwapChain(VulkanContext& context, VkSurfaceKHR vksurface);
+    VulkanSwapChain(VulkanContext& context, uint32_t width, uint32_t height);
+
+    bool acquire();
+    void create();
+    void destroy();
+    void makePresentable();
+
+    VulkanContext& context;
+    VkSurfaceKHR surface;
+    VkSwapchainKHR swapchain;
+    VkSurfaceCapabilitiesKHR surfaceCapabilities;
+    VkSurfaceFormatKHR surfaceFormat;
+    VkExtent2D clientSize;
+    std::vector<VkSurfaceFormatKHR> surfaceFormats;
+    VkQueue presentQueue;
+    VkQueue headlessQueue;
+    std::vector<VulkanAttachment> attachments;
+    uint32_t currentSwapIndex;
+
+    // This is signaled when vkAcquireNextImageKHR succeeds, and is waited on by the first
+    // submission.
+    VkSemaphore imageAvailable;
+
+    // This is true after the swap chain image has been acquired, but before it has been presented.
+    bool acquired;
+
+    VulkanAttachment depth;
+    bool suboptimal;
+    bool firstRenderPass;
+};
+
+} // namespace filament
+} // namespace backend
+
+#endif // TNT_FILAMENT_DRIVER_VULKANTEXTURE_H
